# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Node.js API with JWT Authentication",
    "description": "API for managing Users, Roles, Claims, and TestObjects with JWT-based Authentication"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "username": { "type": "string", "example": "johndoe" },
          "role": { "type": "string", "example": "admin" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Unauthorized" },
          "status": { "type": "integer", "example": 401 }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "description": "Retrieve a list of all users (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "description": "Create a new user (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string" }
                },
                "required": ["username", "password", "role"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a user by ID",
        "description": "Retrieve a user by their ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "User not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user by ID",
        "description": "Update a user by their ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "role": { "type": "string" }
                },
                "required": ["username", "role"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "User not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user by ID",
        "description": "Delete a user by their ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": { "description": "User deleted successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "User not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    },
...
  }
}